% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoTools.R
\name{aggregateArray}
\alias{aggregateArray}
\title{Aggregate a array(dim >=3)* object to create a lower resolution (larger cells).
Aggregation groups rectangular areas to create larger cells.
The value for the resulting cells is computed with a user-specified function.}
\usage{
aggregateArray(OriginalWert, aggregateN, aggregateFunction = colMeans)
}
\arguments{
\item{OriginalWert}{Original array Wert}

\item{aggregateN}{integer. Aggregation factor expressed as number of cells in each direction (horizontally and vertically).
Or two integers (horizontal and vertical aggregation factor) or three integers (when also aggregating over layers).}

\item{aggregateFunction}{= colMeans, function used to aggregate values. Function must support vectorization of column as input, like base::colMeans.}
}
\value{
a lower resolution array(same dimensions)
}
\description{
Aggregate a array(dim >=3)* object to create a lower resolution (larger cells).
Aggregation groups rectangular areas to create larger cells.
The value for the resulting cells is computed with a user-specified function.
}
