% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckTools.R
\name{checkData}
\alias{checkData}
\title{checkData}
\usage{
checkData(Aim, Check, name)
}
\arguments{
\item{Aim}{a variable in list}

\item{Check}{the will be checkd variable}

\item{name}{the name of the variable}
}
\value{
NULL, the message will on Console print
}
\description{
check the structure and names of the elements
}
