% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bel_Aneich_Funct.R
\name{delete_BS_A}
\alias{delete_BS_A}
\title{loesen die Bodenstationen}
\usage{
delete_BS_A(Rd_Data, BS_Data, pixeln_Weit = 1, Schwellewert = 0.1)
}
\arguments{
\item{Rd_Data}{Radar Daten in raster.}

\item{BS_Data}{Bodenstation Daten in data.frame mit(am mindesten) "Lat", "Lon", "Value".}

\item{pixeln_Weit}{Pixeln weit vom Such, 1 = 9-Pixeln, 2 = 25-Pixeln des Umfangs.}

\item{Schwellewert}{Schwellewert fuer loesen.}
}
\value{
neu Daten in data.frame
}
\description{
loesen die Bodenstationen
}
