% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hm_listTools.R
\name{delete_sub_list}
\alias{delete_sub_list}
\title{delete_sub_list}
\usage{
delete_sub_list(Ori, New)
}
\arguments{
\item{Ori}{the original data}

\item{New}{the new data}
}
\value{
a data deleted from Ori and New
}
\description{
delete the data from Ori and New bei elments names
}
