% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListTools.R
\name{deleteData.data.frame}
\alias{deleteData.data.frame}
\title{deleteData.data.frame}
\usage{
\method{deleteData}{data.frame}(Left, Right)
}
\arguments{
\item{Left}{the left data}

\item{Right}{the right data}
}
\value{
a data deleted from Left
}
\description{
delete the data in Left, which ist(are) auch in Right bei elments names for data.frame
}
