% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_convolution.R
\name{discrete_convolution}
\alias{discrete_convolution}
\title{discrete convolution}
\usage{
discrete_convolution(VectorSmall, VectorBig)
}
\arguments{
\item{VectorSmall}{the Small Vector}

\item{VectorBig}{the Big Vector}
}
\value{
wert of discrete convolution
}
\description{
discrete convolution
}
