% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bel_Aneich_Funct.R
\name{faktor_BS}
\alias{faktor_BS}
\title{Faktoren oder Differenz in Bodenstationen}
\usage{
faktor_BS(Rd_Data, BS_Data, methond = "A", pixeln_Weit = 1)
}
\arguments{
\item{Rd_Data}{Radar Daten in raster.}

\item{BS_Data}{Bodenstation Daten in data.frame mit(am mindesten) "Lat", "Lon", "Value".}

\item{methond}{"A" oder "D", A = Faktorenverfahren, D = Differenyverfahren.}

\item{pixeln_Weit}{Pixeln weit vom Such, 1 = 9-Pixeln, 2 = 25-Pixeln des Umfangs.}
}
\value{
Faktoren der Bodenstationen in data.frame
}
\description{
Faktoren oder Differenz in Bodenstationen
}
