% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{fctClassify}
\alias{fctClassify}
\title{High resolution data reclassfy to low resolution data: N*N data to 1 data}
\usage{
fctClassify(filePath, GridID, classifyN, leftDropN = 0, upDropN = 0)
}
\arguments{
\item{filePath}{filename}

\item{GridID}{2-array, the gridID in geological}

\item{classifyN}{intger how many data to 1 data}

\item{leftDropN}{how many grid in left in High resolution data would deleate}

\item{upDropN}{how many grid in up in High resolution data would deleate}
}
\value{
a new data mit low resolution data
}
\description{
High resolution data reclassfy to low resolution data: N*N data to 1 data
}
