% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoTools.R
\name{fctFindNext}
\alias{fctFindNext}
\title{find the netx grid ID starting at a given point.}
\usage{
fctFindNext(x, p, ...)
}
\arguments{
\item{x}{RasterLayer of flow direction (as can be created with terrain}

\item{p}{starting point. Either two numbers: x (longitude) and y (latitude) coordinates; or a single cell number}

\item{...}{other parameter}
}
\value{
numeric (start ID and next ID)
}
\description{
find the netx grid ID starting at a given point.
}
