% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoTools.R
\name{fctGridArea}
\alias{fctGridArea}
\title{caculate area for every grids}
\usage{
fctGridArea(BasinRaster, IDGrid, classifyN)
}
\arguments{
\item{BasinRaster}{RasterLayer, values Are 1}

\item{IDGrid}{RasterLayer, values are ID}

\item{classifyN}{numica, classsify N}
}
\value{
Grid Area in RasterLayer
}
\description{
caculate area for every grids
}
