% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Station2GridIDW.R
\name{fctlengthtem}
\alias{fctlengthtem}
\title{caculate abstand}
\usage{
fctlengthtem(a, b)
}
\arguments{
\item{a}{n* 2(latiyude, longitude) matix of station}

\item{b}{n * 2(latiyude, longitude) 2-array of Grid}
}
\value{
c
}
\description{
caculate abstand
}
