% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTools.R
\name{fctMon2Period}
\alias{fctMon2Period}
\title{Make the month Data to the period Data.}
\usage{
fctMon2Period(NMon, MonData)
}
\arguments{
\item{NMon}{the moth in number}

\item{MonData}{the data in month scaler(12 \* gridN) or (gridN \* 12)}
}
\value{
the data in periodN
}
\description{
Make the month Data to the period Data.
}
