% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_sp.R
\name{geom_basin}
\alias{geom_basin}
\alias{geom_river}
\title{translate raster(RasterLayer) to data.frame, inoder to use ggplot}
\usage{
geom_basin(sp_basin, Color = "#e31a1c", Size = 1.3, ...)

geom_river(sp_river, Color = "#6baed6", Size = 1, ...)
}
\arguments{
\item{sp_basin}{SpatialPolygonsDataFrame, basin}

\item{Color}{color for linie in geom_path}

\item{Size}{size for linie in geom_path}

\item{...}{other parameters for geom_path}

\item{sp_river}{SpatialPolygonsDataFrame, river}
}
\value{
data.frame, inoder to use ggplot.
}
\description{
translate raster(RasterLayer) to data.frame, inoder to use ggplot
}
