% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_sp.R
\name{geom_dir}
\alias{geom_dir}
\title{new funtion for geom_raster in ggplot2}
\usage{
geom_dir(
  DirGrid,
  dirColour = "#084594",
  dirArrow = arrow(length = unit(0.2, "cm")),
  ...
)
}
\arguments{
\item{DirGrid}{RasterLayer,same size with DirGrid}

\item{dirColour}{colour in geom_segment}

\item{dirArrow}{arrow in geom_segment}

\item{...}{other paramters for geom_}
}
\value{
ggploto like geo_raster
}
\description{
new funtion for geom_raster in ggplot2
}
