% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_sp.R
\name{geom_raster.seq}
\alias{geom_raster.seq}
\alias{geom_raster.div}
\alias{geom_raster.qual}
\title{new funtion for geom_ in ggplot2}
\usage{
geom_raster.seq(raster, ...)

geom_raster.div(raster, breaks, ...)

geom_raster.qual(raster, ...)
}
\arguments{
\item{raster}{RasterLayer}

\item{...}{other paramters for geom_}

\item{breaks}{breakslikes in cut}
}
\value{
ggploto like geo_raster
}
\description{
new funtion for geom_ in ggplot2
}
