% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListTools.R
\name{getI}
\alias{getI}
\title{getI}
\usage{
getI(Data, i)
}
\arguments{
\item{Data}{Data}

\item{i}{index}
}
\value{
the ith data of one demision data or get the ith row of two demisions data
}
\description{
get the ith data of one demision data or get the ith row of two demisions data
}
