% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hm_listTools.R
\name{getI_sub_hm.list}
\alias{getI_sub_hm.list}
\title{getI_sub_hm.list}
\usage{
getI_sub_hm.list(Data, i, value)

getI_sub_hm.list(Data, i, value)
}
\arguments{
\item{Data}{the data, from it get the required data}

\item{i}{index}

\item{value}{PutData}
}
\value{
Locat whith the required data

list
}
\description{
Extract the data required by Locat in Data for list

put the data to a HM.Data
}
