% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hm_listTools.R
\name{getI_t_vari.array}
\alias{getI_t_vari.array}
\title{subsetting for HM.Data}
\usage{
getI_t_vari.array(Data, i)
}
\arguments{
\item{Data}{array in class t_vari.array}

\item{i}{index}
}
\value{
Locat whith the required data
}
\description{
Extract the data
}
