% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmlist.R
\name{getListI}
\alias{getListI}
\title{getListI}
\usage{
getListI(List, i)
}
\arguments{
\item{List}{the data, from it get the required data}

\item{i}{index}
}
\value{
Locat whith the required data
}
\description{
Extract the data required by Locat in Data for list
}
