% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_sp.R
\name{ggplot_dem}
\alias{ggplot_dem}
\title{new funtion for geom_raster in ggplot2}
\usage{
ggplot_dem(DEM)
}
\arguments{
\item{DEM}{RasterLayer, dem}
}
\value{
ggplot like ggplot()
}
\description{
new funtion for geom_raster in ggplot2
}
