% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GID}
\alias{GID}
\title{grid-data of gridID}
\format{
'data.frame' or 2-array
\describe{
mit attributs
\item{NOdata}{is NA}
}
}
\usage{
GID
}
\description{
grid-data of gridID
}
\keyword{datasets}
