% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLC}
\alias{GLC}
\title{location infomation of grids.}
\format{
'data.frame':	gridN obs. of  3 variables
\describe{
(The field names in these lists must be consistent and besause they are the only index.)
\item{ID}{id code of grids}
\item{Latitude}{Latitude of grids}
\item{Longitude}{Longitude of grids}
}
}
\usage{
GLC
}
\description{
location infomation of grids.
}
\keyword{datasets}
