% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoTools.R
\name{gridParamJion}
\alias{gridParamJion}
\title{read grid-data from ArcGIS}
\usage{
gridParamJion(GridCodeTable, gridN, ParamLib)
}
\arguments{
\item{GridCodeTable}{datafram mit GridID, wert, Rate. name of Wert must same with first column of ParamLib.(output from fctClassfy)}

\item{gridN}{intger, how many grid.}

\item{ParamLib}{datafram, Data set corresponding to each type.}
}
\value{
a datafram(gridN, paramN)
}
\description{
read grid-data from ArcGIS
}
\examples{
GL <- fctClassify(OGL, GID, 20, 0, 5)
names(GL)[2] <- "Code"
GLP <- gridParamJion(GL, 311, LLib)
}
