% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListTools.R
\name{interjoinData}
\alias{interjoinData}
\title{interjoinData}
\usage{
interjoinData(Left, Right)
}
\arguments{
\item{Left}{the left data}

\item{Right}{the right data}
}
\value{
a data innerjoind from Left and Right
}
\description{
innerjoin the data from Left and Right bei elments names
}
