% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bel_Aneich_Funct.R
\name{interpolat_AF}
\alias{interpolat_AF}
\title{Interpolation fuer  oder Differenz in Bodenstationen}
\usage{
interpolat_AF(Mat_g, AF_BS)
}
\arguments{
\item{Mat_g}{Output aus "radar_Gewicht".}

\item{AF_BS}{Output aus Funktion "faktor_BS"}
}
\value{
Interpolatiert Faktoren.
}
\description{
Interpolation fuer  oder Differenz in Bodenstationen
}
