% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bel_Aneich_Funct.R
\name{interpolat_Kriging}
\alias{interpolat_Kriging}
\title{Interpolation mit original Kriging}
\usage{
interpolat_Kriging(BS_Data, Rd_Data, cacuStationN = 5, sAnstgVario = 4.25)
}
\arguments{
\item{BS_Data}{Bodenstation Daten in data.frame mit(am mindesten) "Lat", "Lon", "Value".}

\item{Rd_Data}{Radar Daten in raster.}

\item{cacuStationN}{wie viele Station zu berücksichtigen.}

\item{sAnstgVario}{Parameter für Kringing.}
}
\value{
Interpolatiert Value.
}
\description{
Interpolation mit original Kriging
}
