% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{interpolation}
\alias{interpolation}
\title{caculate interpolation x in X, find y in Y}
\usage{
interpolation(bktprm, BKTPRMLST, ZLPRMLST)
}
\arguments{
\item{bktprm}{input x}

\item{BKTPRMLST}{X}

\item{ZLPRMLST}{Y}
}
\value{
y
}
\description{
caculate interpolation x in X, find y in Y
}
\examples{
interpolation(2.5, 1:3, c(2, 8, 13))
}
