% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hm_listTools.R
\name{intersect_sub_list}
\alias{intersect_sub_list}
\title{intersect_sub_list}
\usage{
intersect_sub_list(Ori, New)
}
\arguments{
\item{Ori}{the original data}

\item{New}{the new data}
}
\value{
a data intersectd from Ori and New
}
\description{
intersect the data from Ori and New bei elments names
}
