% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hm_listTools.R
\name{join_sub_list}
\alias{join_sub_list}
\title{join_sub_list}
\usage{
join_sub_list(Ori, New)
}
\arguments{
\item{Ori}{the original data}

\item{New}{the new data}
}
\value{
a data joind from Ori and New
}
\description{
join the data from Ori and New bei elments names
}
