% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hm_listTools.R
\name{left_delete.hm.list}
\alias{left_delete.hm.list}
\title{left_delete}
\usage{
\method{left_delete}{hm.list}(Ori, New)
}
\arguments{
\item{Ori}{the original data}

\item{New}{the new data}
}
\value{
a data deleted from Ori and New
}
\description{
delete the data from Ori and New bei elments names
}
