% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hm_listTools.R
\name{left_delete}
\alias{left_delete}
\title{left_delete}
\usage{
left_delete(Ori, New)
}
\arguments{
\item{Ori}{the original data}

\item{New}{the new data}
}
\value{
a data deleted from Ori and New
}
\description{
delete the data from Ori and New bei elments names
}
