% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hm_listTools.R
\name{left_intersect.hm.list}
\alias{left_intersect.hm.list}
\title{left_intersect}
\usage{
\method{left_intersect}{hm.list}(Ori, New)
}
\arguments{
\item{Ori}{the original data}

\item{New}{the new data}
}
\value{
a data intersectd from Ori and New
}
\description{
intersect the data from Ori and New bei elments names
}
