% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hm_listTools.R
\name{left_join.hm.list}
\alias{left_join.hm.list}
\title{left_join}
\usage{
\method{left_join}{hm.list}(Ori, New)
}
\arguments{
\item{Ori}{the original data}

\item{New}{the new data}
}
\value{
a data joind from Ori and New
}
\description{
join the data from Ori and New bei elments names
}
