% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hm_listTools.R
\name{left_join}
\alias{left_join}
\title{left_join}
\usage{
left_join(Ori, New)
}
\arguments{
\item{Ori}{the original data}

\item{New}{the new data}
}
\value{
a data joind from Ori and New
}
\description{
join the data from Ori and New bei elments names
}
