% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListTools.R
\name{left_merge.list}
\alias{left_merge.list}
\title{merge_sub_list}
\usage{
\method{left_merge}{list}(Ori, New)
}
\arguments{
\item{Ori}{the original data}

\item{New}{the new data}
}
\value{
a data merged from Ori and New
}
\description{
merge the data from Ori and New bei elments names
}
