% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hm_listTools.R
\name{left_merge}
\alias{left_merge}
\title{mergeData}
\usage{
left_merge(Ori, New)
}
\arguments{
\item{Ori}{the original data}

\item{New}{the new data}
}
\value{
a data merged from Ori and New
}
\description{
merge the data from Ori and New bei elments names
}
