% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListTools.R
\name{leftjoinData}
\alias{leftjoinData}
\title{leftjoinData}
\usage{
leftjoinData(Left, Right)
}
\arguments{
\item{Left}{the left data}

\item{Right}{the right data}
}
\value{
a data leftjoind from Left and Right
}
\description{
leftjoin the data from Left and Right bei elments names
}
