% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Modul_Tools.R
\name{merge_Data_Modul}
\alias{merge_Data_Modul}
\title{merge the Data for every Modules}
\usage{
merge_Data_Modul(...)
}
\arguments{
\item{...}{the Data for Modul A B C and so on}
}
\value{
a data merged from Data_Model.A B C and so on
}
\description{
merge the data
}
