% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Modul_Tools.R
\name{merge_mdl_data2}
\alias{merge_mdl_data2}
\title{merge the Data for every Modules}
\usage{
merge_mdl_data2(Data_Model.A, Data_Model.B)
}
\arguments{
\item{Data_Model.A}{the Data for Modul A}

\item{Data_Model.B}{the Data for Modul B}
}
\value{
a data merged from Data_Model.A and Data_Model.A
}
\description{
merge the data
}
