% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hm_listTools.R
\name{merge_sub_list}
\alias{merge_sub_list}
\title{merge_sub_list}
\usage{
merge_sub_list(Ori, New)
}
\arguments{
\item{Ori}{the original data}

\item{New}{the new data}
}
\value{
a data merged from Ori and New
}
\description{
merge the data from Ori and New bei elments names
}
