% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMDataTools.R
\name{mergeClass}
\alias{mergeClass}
\title{mergeData.list}
\usage{
mergeClass(cName, Ori, New)
}
\arguments{
\item{cName}{class name}

\item{Ori}{the original data}

\item{New}{the new data}
}
\value{
a data merged from Ori and New
}
\description{
merge the data from Ori and New bei elments names for list
}
