% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListTools.R
\name{mergeData.data.frame}
\alias{mergeData.data.frame}
\title{mergeData.data.frame}
\usage{
\method{mergeData}{data.frame}(Ori, New)
}
\arguments{
\item{Ori}{the original data}

\item{New}{the new data}
}
\value{
a data merged from Ori and New
}
\description{
merge the data from Ori and New bei elments names for data.frame
}
