% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMData.R, R/HMDataTools.R
\name{mergeData.HM.Data}
\alias{mergeData.HM.Data}
\title{mergeData}
\usage{
\method{mergeData}{HM.Data}(Ori, New)

\method{mergeData}{HM.Data}(Ori, New)
}
\arguments{
\item{Ori}{the original data}

\item{New}{the new data}
}
\value{
a data merged from Ori and New

a data merged from Ori and New
}
\description{
merge the data from Ori and New bei elments names

merge the data from Ori and New bei elments names for list
}
