% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmlist.R
\name{mergeData.hm.list}
\alias{mergeData.hm.list}
\title{mergeData}
\usage{
\method{mergeData}{hm.list}(Ori, New)
}
\arguments{
\item{Ori}{the original data}

\item{New}{the new data}
}
\value{
a data merged from Ori and New
}
\description{
merge the data from Ori and New bei elments names
}
