% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMDataTools.R
\name{mergeSlots}
\alias{mergeSlots}
\title{mergeData.list}
\usage{
mergeSlots(Ori, New)
}
\arguments{
\item{Ori}{the original data}

\item{New}{the new data}
}
\value{
a data merged from Ori and New
}
\description{
merge the data from Ori and New bei elments names for list
}
