% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hm_listTools.R
\name{na_check}
\alias{na_check}
\title{na_check}
\usage{
na_check(Data)
}
\arguments{
\item{Data}{the original data}
}
\value{
a data deleted from Ori and New
}
\description{
check the NA in the hm.list data
}
