% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OGL}
\alias{OGL}
\title{grid-data of landuse.}
\format{
'data.frame' or 2-array
\describe{
mit attributs
\item{NOdata}{is NA}
}
}
\usage{
OGL
}
\description{
grid-data of landuse.
}
\keyword{datasets}
