% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListTools.R
\name{putAttr}
\alias{putAttr}
\title{put the atrribut}
\usage{
putAttr(Data, index, AttrNa, AttrVa)
}
\arguments{
\item{Data}{Data of list or data,frame}

\item{index}{index for the Data}

\item{AttrNa}{name of the attribut}

\item{AttrVa}{value of atrribut}
}
\value{
data with atrribut
}
\description{
put the atrribut
}
