% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListTools.R
\name{putData.default}
\alias{putData.default}
\title{putData.default}
\usage{
\method{putData}{default}(Data, Locat, i)
}
\arguments{
\item{Data}{the data, from it get the required data}

\item{Locat}{the loction, the data will in it put}

\item{i}{index}
}
\value{
Locat whith the required data
}
\description{
put the data required by Locat in Data into Locat for other datda, exp. matrix or array
}
