% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListTools.R
\name{putI}
\alias{putI}
\title{putI}
\usage{
putI(Data, i, PutD)
}
\arguments{
\item{Data}{Data}

\item{i}{index}

\item{PutD}{Put Data}
}
\value{
put the ith data of one demision data or put the ith row of two demisions data
}
\description{
get the ith data of one demision data or get the ith row of two demisions data
}
