% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hm_listTools.R
\name{putI_sub_hm.list}
\alias{putI_sub_hm.list}
\title{putList}
\usage{
putI_sub_hm.list(Data, i, value)
}
\arguments{
\item{Data}{the data, from it get the required data}

\item{i}{index}

\item{value}{PutData}
}
\value{
list
}
\description{
put the data to a HM.Data
}
