% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMData.R
\name{putList}
\alias{putList}
\title{putList}
\usage{
putList(Data, i, PutData)
}
\arguments{
\item{Data}{the data, from it get the required data}

\item{i}{index}

\item{PutData}{PutData}
}
\value{
list
}
\description{
put the data to a HM.Data
}
