% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListTools.R
\name{putUnit}
\alias{putUnit}
\title{put the unit atrribut}
\usage{
putUnit(Data, AttrVa)
}
\arguments{
\item{Data}{Data of list or data,frame}

\item{AttrVa}{value of unit}
}
\value{
data with unit
}
\description{
put the unit atrribut
}
