% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bel_Aneich_Funct.R
\name{radar_Gewicht}
\alias{radar_Gewicht}
\title{Faktoren oder Differenz in Bodenstationen}
\usage{
radar_Gewicht(BS_lon_lat, Rd_Raster, Such_radius = 1250)
}
\arguments{
\item{BS_lon_lat}{Bodenstation Daten in data.frame nur mit "Lon" und "Lat".}

\item{Rd_Raster}{Radar Daten in raster.}

\item{Such_radius}{Wie weit sucht man die Bodenstation zu interpolatieren.}
}
\value{
Gewichten füer jeden Pixel.
}
\description{
Faktoren oder Differenz in Bodenstationen
}
