% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_sp.R
\name{raster.dir}
\alias{raster.dir}
\title{translate raster(RasterLayer) to data.frame, inoder to use ggplot}
\usage{
raster.dir(DirGrid)
}
\arguments{
\item{DirGrid}{RasterLayer, dirction from terrn()}
}
\value{
data.frame, inoder to use ggplot.
}
\description{
translate raster(RasterLayer) to data.frame, inoder to use ggplot
}
