% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_sp.R
\name{raster.seq}
\alias{raster.seq}
\alias{raster.div}
\alias{raster.qual}
\title{translate raster(RasterLayer) to data.frame, inoder to use ggplot}
\usage{
raster.seq(raster)

raster.div(raster, breaks)

raster.qual(raster)
}
\arguments{
\item{raster}{RasterLayer}

\item{breaks}{breakslikes in cut}
}
\value{
data.frame, inoder to use ggplot.
}
\description{
translate raster(RasterLayer) to data.frame, inoder to use ggplot
}
