% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoTools.R
\name{read.grid}
\alias{read.grid}
\title{read grid-data from ArcGIS}
\usage{
read.grid(filePath, ...)
}
\arguments{
\item{filePath}{filename grid-data aus ArcGIS, there are 6 row ist meta data}

\item{...}{other paramter}
}
\value{
a datafaram mit 6 attributs
}
\description{
read grid-data from ArcGIS
}
