% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SD}
\alias{SD}
\title{metrol infomation from hydro station.}
\format{
3-array(periodN, stationN, fieldN) of data from station
}
\usage{
SD
}
\description{
metrol infomation from hydro station.
}
\keyword{datasets}
