% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SLC}
\alias{SLC}
\title{location infomation of hydro station.}
\format{
'data.frame':	stationN obs. of  3 variables
\describe{
(The field names in these lists must be consistent and besause they are the only index.)
\item{ID}{id code of station}
\item{Latitude}{Latitude of station}
\item{Longitude}{Longitude of station}
}
}
\usage{
SLC
}
\description{
location infomation of hydro station.
}
\keyword{datasets}
