% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_sp.R
\name{sp.path}
\alias{sp.path}
\title{translate raster(RasterLayer) to data.frame, inoder to use ggplot}
\usage{
sp.path(sppath)
}
\arguments{
\item{sppath}{SpatialPolygonsDataFrame}
}
\value{
data.frame, inoder to use ggplot.
}
\description{
translate raster(RasterLayer) to data.frame, inoder to use ggplot
}
