% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Station2GridKrige.R
\name{Station2GridKrige}
\alias{Station2GridKrige}
\title{Station-Data interpolation to Grid-Data with point-Krige}
\usage{
Station2GridKrige(
  StationLocation,
  GridLocation,
  StationData,
  cacuStationN = 5,
  wightCoefficient = 2,
  sAnstgVario = 4.25
)
}
\arguments{
\item{StationLocation}{n* 3(ID, Latitude, Longitude) 2-array of station}

\item{GridLocation}{n * 3(ID, Latitude, Longitude) 2-array of Grid}

\item{StationData}{3-array(periodN, stationN, fieldN) of data from station}

\item{cacuStationN}{= 5 how many Station are relevant for a grid}

\item{wightCoefficient}{= 2 wight Coefficient k, 1-5}

\item{sAnstgVario}{= 4.25 linearen Variogrammes s}
}
\value{
Ntime * Nstation * Nfild 3-array of data for Grids
}
\description{
Station-Data interpolation to Grid-Data with point-Krige
}
\examples{
GridMetroData <- Station2GridKrige(SLC, GLC, SD)
}
