% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMData.R
\name{[.HM.Data}
\alias{[.HM.Data}
\title{subsetting for HM.Data}
\usage{
\method{[}{HM.Data}(Data, i)
}
\arguments{
\item{Data}{the data, from it get the required data}

\item{i}{index}
}
\value{
Locat whith the required data
}
\description{
Extract the data
}
