% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmlist.R
\name{[.hm.list}
\alias{[.hm.list}
\title{subsetting for hm.list}
\usage{
\method{[}{hm.list}(Data, i)
}
\arguments{
\item{Data}{the data, from it get the required data}

\item{i}{index}
}
\value{
Locat whith the required data
}
\description{
Extract the data
}
