% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMData.R
\name{[<-.HM.Data}
\alias{[<-.HM.Data}
\title{subsetting geben for HM.Data}
\usage{
\method{[}{HM.Data}(Data, i) <- value
}
\arguments{
\item{Data}{the data, from it get the required data}

\item{i}{index}

\item{value}{value}
}
\value{
new data
}
\description{
Extract the data required by Locat in Data for list
}
