% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hm_listTools.R
\name{[<-.t_vari.hm.list}
\alias{[<-.t_vari.hm.list}
\title{subsetting geben for HM.Data}
\usage{
\method{[}{t_vari.hm.list}(Data, i) <- value
}
\arguments{
\item{Data}{the data, from it get the required data}

\item{i}{index}

\item{value}{value}
}
\value{
new data
}
\description{
Extract the data required by Locat in Data for list
}
