% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTools.R
\name{toMON}
\alias{toMON}
\title{caculate month in upper abb}
\usage{
toMON(Date)
}
\arguments{
\item{Date}{Date or date list}
}
\value{
ABB month in upper abb
}
\description{
caculate month in upper abb
}
\examples{
toMON("2020-4-1")
}
