% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTools.R
\name{toNDayofYear}
\alias{toNDayofYear}
\title{caculate which day of this year}
\usage{
toNDayofYear(DateDay)
}
\arguments{
\item{DateDay}{nur Date or vector of Date}
}
\value{
a one-dimesion array
}
\description{
caculate which day of this year
}
\examples{
toNDayofYear(as.Date("2020-2-23"))
toNDayofYear(seq(as.Date("1996-2-23"),as.Date("2020-2-23"),1))
}
