% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckTools.R
\name{viewArgum}
\alias{viewArgum}
\title{viewArgum}
\usage{
viewArgum(fcName, Arguments)
}
\arguments{
\item{fcName}{function name}

\item{Arguments}{Arguments}
}
\value{
return a list, includ Arguments and outdata,
they have the same structure whith Arguments and outdata, but not the real werts
}
\description{
view the Arguments of a function
}
