% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckTools.R
\name{viewData}
\alias{viewData}
\title{viewData}
\usage{
viewData(Model, name)
}
\arguments{
\item{Model}{a variable in list}

\item{name}{the name of the variable}
}
\value{
NULL, the message will on Console print
}
\description{
view the structure and names of the elements
}
