% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckTools.R
\name{viewFuPa}
\alias{viewFuPa}
\title{viewFuPa}
\usage{
viewFuPa(Name, clsNa)
}
\arguments{
\item{Name}{the Name of funtions oder parameters}

\item{clsNa}{the class name of S3 function}
}
\value{
for function return the Input data and Out data structure
for paramter return the wert and description
}
\description{
view the funtions oder parameters
}
