% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoTools.R
\name{fctNextGrid}
\alias{fctNextGrid}
\title{caculate next grid from DirGrid}
\usage{
fctNextGrid(IDGrid, DirGrid)
}
\arguments{
\item{IDGrid}{RasterLayer,same size with DirGrid}

\item{DirGrid}{RasterLayer,from raster::terrain(dem, opt = "flowdir")}
}
\value{
RasterLayer, next grid ID
}
\description{
caculate next grid from DirGrid
}
