% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoTools.R
\name{fctNextGrid}
\alias{fctNextGrid}
\title{caculate ID of Next Grid}
\usage{
fctNextGrid(DirGrid, IDGrid)
}
\arguments{
\item{DirGrid}{RasterLayer,from raster::terrain(dem, opt = "flowdir")}

\item{IDGrid}{RasterLayer,same size with DirGrid}
}
\value{
2-array(gridN,2) gridID and Next Grid ID
}
\description{
caculate ID of Next Grid
}
