% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmlist.R
\name{putList}
\alias{putList}
\title{putData.list}
\usage{
putList(PutData, Data, i)
}
\arguments{
\item{PutData}{PutData}

\item{Data}{the data, from it get the required data}

\item{i}{index}
}
\value{
Locat whith the required data
}
\description{
Extract the data required by Locat in Data for list
}
