% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Station2GridIDW.R
\name{Station2GridIDW}
\alias{Station2GridIDW}
\title{Station-Data interpolation to Grid-Data with IDW}
\usage{
Station2GridIDW(StationLocation, GridLocation, StationData, StationN = 5)
}
\arguments{
\item{StationLocation}{n* 2(latiyude, longitude) matix of station}

\item{GridLocation}{n * 2(latiyude, longitude) 2-array of Grid}

\item{StationData}{Ntime * Nstation * Nfild 3-array of data from station}

\item{StationN}{how many Station are relevant for a grid}
}
\value{
Ntime * Nstation * Nfild 3-array of data for Grids
}
\description{
Station-Data interpolation to Grid-Data with IDW
}
