% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmlist.R
\name{[<-.hm.list}
\alias{[<-.hm.list}
\title{subsetting geben for hm.list}
\usage{
\method{[}{hm.list}(Data, i) <- value
}
\arguments{
\item{Data}{the data, from it get the required data}

\item{i}{index}

\item{value}{value}
}
\value{
new data
}
\description{
Extract the data required by Locat in Data for list
}
